+++ src/main/Java/dukeychatbot/dukeyexceptions/MissingTimeframeException.java
@@ -1,9 +1,15 @@
-package Exceptions;
+package dukeychatbot.dukeyexceptions;
 
-public class MissingTimeframeException extends DukeException {
+/**
+ * Constructs MissingTimeframeException error which inherits from DukeyException.
+ *
+ * <p>Returns error when user fails to provide timeframe for Event tasks.</p>
+ */
+public class MissingTimeframeException extends DukeyException {
+
     public MissingTimeframeException() {
         super("____________________________________________________________\n" +
-                "WARNING: Event command requires a timeframe.\n" +
+                "WARNING: dukeychatbot.tasktypes.Event command requires a timeframe.\n" +
                 "Valid input requires '/from' AND '/to' keyword. " +
                 "Follow the format: event <task name> /from <date / timing> /to <date / timing> \n" +
                 "E.g. event project meeting /from Mon 2pm /to 4pm\n" +
