--- src/main/java/Exceptions/DukeException.java
@@ -1,8 +1,0 @@
-package Exceptions;
-
-public class DukeException extends Exception {
-
-    public DukeException(String message) {
-        super(message);
-    }
-}
