--- src/main/java/Test.java
@@ -1,4 +1,0 @@
-package PACKAGE_NAME;
-
-public class Test {
-}
